+++ cmd/cosign/cli/fulcio/fulcioroots/fulcio.pem
